feat: unify sales home entry
